refactor: removed version prefix from routers
