--- app/routers/v1/fhir/__init__.py
@@ -1,3 +1,0 @@
-SUBJECT_IDENTIFIER_PARAM = "subject:identifier"
-DEVICE_IDENTIFIER_PARAM = "source:identifier"
-CODE_PARAM = "code"
